renamed to config for better understanding
